Remove workloads, reorganize hooks, fix mission namepsace bug
